cobalt/kernel: Switch back ptraced primary mode threads on resume
When a thread is stopped in primary mode for debugging, make sure it
will migrate back before resuming in user space. This is a building
block for making real-time process debugging more deterministic.

The information if a thread should resume in primary is transported via
a new thread info flag XNCONTHI. It is set either by the exception
handler detecting a breakpoint hit in primary user mode or by
xnthread_relax when invoked for a thread under XNSSTEP which indicates
it was signaled to stop.

The feature depends on the new I-pipe notifier for user interrupt
return, i.e. a callback from the point that a secondary mode thread is
about to return to user space after an exception or an interrupt.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>