cobalt/kernel: Implement synchronous stop and resume for process debugging
When a thread in primary mode currently hits a breakpoint, it will
immediately be relaxed, leaving space for other runnable RT threads of
the process to take over. This can delay gdb to finally take control and
stop the whole process.

To prevent this unexpected and often undesired behavior, this introduces
a mechanism to stop all threads of a process in primary mode when one of
them runs into a breakpoint or a single-step event. The new XNDBGSTOP
thread state is used for this purpose.

On resumption, no thread can continue in primary mode until all that
should do so have reached this state. Then the last one will wake up all
threads waiting on XNDBGSTOP.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>