cobalt/kernel: Rework register/unregister_debugged_thread
Given that registration can run asynchronously to deregistration, better
put testing for XNSSTEP and the latter under nklock. This is likely
overkill for task termination but it's more consistent.

While at it, move up the functions and add a cobalt_ prefix. We will
need them earlier in the file and are going to export one of them for
use in other code modules.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>