net/tcp: fix TCP connection termination
Closing a connected TCP socket initiates a handshake procedure. The
closing socket is supposed to receive two more packets and send one
ACK at the end. RTnet's TCP stack tries to get a reference to a
socket before writing data to its buffer. Taking a reference to an
RTnet socket in Xenomai 3 is realized by calling rtdm_fd_lock() which
fails when a socket is about to be closed.

Signed-off-by: Sebastian Smolorz <sebastian.smolorz@gmx.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>