drivers/serial: 16550A: Plug race between TX interrupt handler and write function
If the TX interrupt was enabled and fired on empty THR before the write
function updated out_npend, we could get stuck with transmitting. Plug
that race by checking in the write handler if we could write to the FIFO
directly. That ensures early filling of the FIFO as well as that a TX
interrupt is fired again if there needs to be more data written.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>