posix: Make CONFIG_XENO_OPT_DEBUG_POSIX_SYNCHRO widely unconditional
This converts the consistency checks for resource ownership from
optional into unconditional. The motivation is preventing the risk of
kernel crash in the presence of application bugs and the absence of
CONFIG_XENO_OPT_DEBUG_POSIX_SYNCHRO which used to control this check.

The only remaining features that this config manages is then printing a
warning on mutex deadlocks. This is simply moved under the parent
CONFIG_XENO_OPT_DEBUG_USER.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>