drivers: Remove unneeded realtime_core_enabled checks
It's not a good idea to load a driver module without functionality when
the core is not available. Better fail with an error code so that the
user gets proper feedback.

In several cases, this makes the existing realtime_core_enabled checks
obviously superfluous because rtdm_dev_register already performs it.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>