net/tcp: Account for connection teardown handshake
When closing a TCP connection a handshake procedure is executed between the
peers. The close routine of the rttcp driver did not participate in
detecting the end of this handshake but rather waited one second inside
a close call unconditionally. Especially when peers are directly connected
this is a waste of time which can hurt a lot in some situations.

This patch replaces the msleep(1000) call with waiting for a completion.
The maximum waiting time now can be configured via a module parameter.
The completion is done when the termination handshake has finished.

Signed-off-by: Sebastian Smolorz <sebastian.smolorz@gmx.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>