utils/can: fix error handling, use errno instead of -ret
rtcansend and rtcanrecv test and evaluate the return code of POSIX
functions instead of errno. This bug sneaked in with:

d245d6e utils/can: drop dependencies on obsolete rt_dev API

As we are at it, a few missing "\n" of some error messages have been
added as well.

Signed-off-by: Wolfgang Grandegger <wg@grandegger.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>