testsuite/smokey/posix-cond: Make more robust for execution in VMs
Specifically when running over QEMU on ARM64, there are frequent
failures of this test case on check_sleep("mutex_lock"). Apparently,
the context calling thread_spawn runs while the new thread is not yet
executed. When the latter finally will, the creator already released
the mutex, and the sleep check fails. This ensures that both contexts
are running before starting the measurements.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>