cobalt/posix: Add thread_setschedprio
The approach of using thread_setschedparam_ex with __SCHED_CURRENT
failed for policies with complex parameter sets. Therefore, introduce a
syscall that is fully aware of how to apply only the priority change to
the target thread's scheduling settings.

We need a compat path as well because we are passing down pointers.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>