cobalt: sched-quota: Fix live-updating if quotas, once more
f116e7eb9b56 already tried to enable live updates of a quota group but
got one thing wrong and missed another catch:

First, we must not use group_is_active() to check if the group is
currently consuming time. That has to be the same logic as in
xnsched_quota_pick(), i.e. a check of the currently running thread on
the target sched belongs to the modified group.

The second issue is that, if a group had no budget on refill or consumed
it in this period already, an extension of the budget missed to move
expired threads back to the runqueue.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>