kernel/cobalt/posix: Fix cobalt_event_wait for COBALT_EVENT_ALL case
When we wait for all passed event bits to be set, we need to test for
those bits, not the ones that are currently pending. The current code
fails to detect that the condition is met when both the desired as well
as other events are already pending.

Reported by: Ronny Meeus <ronny.meeus@gmail.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>