cobalt/init: pipeline: abstract pipeline-specific inits
This is the initial step to enabling Dovetail for the Cobalt core,
which requires to introduce an abstraction layer between such core and
the interrupt pipeline interface, either Dovetail or the legacy
I-pipe.

Eventually, the Cobalt implementation which has to interface to the
underlying pipeline should branch off at the following points:

- kernel/cobalt/{ipipe, dovetail}/arch, the arch-specific
  implementation which overwhelmingly depends on the pipeline flavour.

- kernel/cobalt/{ipipe, dovetail}, the generic Cobalt code calling
  pipeline services.

- kernel/cobalt/include/{ipipe, dovetail}, the client glue types and
  definitions pulled into some basic kernel types by the pipeline
  implementation (e.g. the thread_info extension structure).

- include/cobalt/kernel/{ipipe, dovetail}, the generic Cobalt headers
  depending on services and definitions the pipeline provides for.

We start the process by abstracting the machine-level init code which
depends on the pipeline flavour.

No functional change is introduced.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>