cobalt/intr: pipeline: IRQ management code is pipeline-specific
The way we request and manage interrupts depends on the underlying
pipeline interface.

As a matter of fact, Dovetail already deals with most of the logic
implemented by the xnintr layer, such as edge/level shared IRQs, fully
reusing the regular genirq interface for management. IRQ handlers with
Dovetail have regular signatures as well.

For the time being, let's move the entire xnintr layer to the I-pipe
specific section created earlier. We should be able to design the
abstract interface to IRQ management after this layer for the most
part, which we would connect to Dovetail eventually.

No functional change is introduced.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>