cobalt/posix/process: pipeline: abstract kernel event handlers
Although there are significant commonalities between the I-pipe and
Dovetail when it comes to dealing with synchronous kernel events,
there is no strict 1:1 mapping between the two kernel interfaces.

As an initial step, move all the code handling the kernel events to
the I-pipe section. We may exploit commonalities between the I-pipe
and Dovetail in this area as we gradually merge support for the
latter.

No functional change is introduced.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>
[Jan: fixed build issues around pipeline_enable_kevents, kallsyms.h and $(srctree)]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>