cobalt/build: pipeline: select IPIPE layer only if present
We may be able to build against a Dovetail-enabled kernel at some
point, so do not force-enable CONFIG_IPIPE, it might not be there.

At this chance, remove obsolete internal switches and conditions. All
I-pipe implementations depend on the GENERIC_CLOCKEVENTS framework,
and support for the legacy I-pipe V1 API is long gone.

No functional change is introduced.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>