cobalt/kernel: pipeline: add abstract synthetic IRQ API
Add wrappers to create "synthetic IRQs" the I-pipe way (used to be
called "virtual IRQs" there). Those interrupt channels can only be
triggered by software, with per-CPU semantics. We use them to schedule
handlers to be run on the in-band execution stage, meaning "secondary
mode" in the Cobalt jargon.

We don't provide for executing handlers on the out-of-band stage,
because Cobalt does not need this.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>
[Jan: broke up one line in pipeline_create_inband_sirq into 2]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>