cobalt/thread: move inband work descriptions off the stack
Unlike the I-pipe, Dovetail does not copy the work descriptor but
merely hands over the request to the common irq_work() mechanism. We
must guarantee that such descriptor lives in a portion of memory which
won't go stale until the handler has run, which by design can only
happen once the calling out-of-band context unwinds.

Therefore, we have to:

- move the in-band work descriptor used in waking up a relaxing thread
  to this thread's control block.

- allocate a request descriptor from the core heap when issuing a
  request to send an in-band signal to a particular thread. Since we
  might have multiple signals sent concurrently to a single thread,
  having a single request descriptor in the thread control block for
  such purpose is not an option.

- allocate a request descriptor from the core heap when issuing a
  request to wake up the parent task of an emerging Cobalt kthread
  (such task is normally a plain regular Linux one, i.e. it does not
  have any associated Cobalt control block we might use for storing
  the descriptor).

Signed-off-by: Philippe Gerum <rpm@xenomai.org>