cobalt/machine: ipipe: drop timer frequency setting
The only user of the timer frequency setting was the machine-specific
timer calibration handler, which is now gone. Therefore we don't need
to know the timer frequency anymore. As a consequence, we don't care
about overriding its value as determined by the I-pipe either.

Let's drop all of these antiquated bits.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>