cobalt/assert: pipeline: add TODO() marker
As its name suggests, TODO() can be added to the code in order to
highlight a place where some implementation bits are known to be
missing. Turning on CONFIG_XENO_TODO reveals the places where such
markers might linger by triggering a build time assertion, so that
they don't go unnoticed.

We may drop this helper when the port to Dovetail is complete.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>