cobalt/clock: dovetail: abstract interface to hardware TSC
Unlike the I-pipe, Dovetail does not need to expose the underlying
timestamp counter to applications. The best equivalent to this counter
would be the common monotonic clock, expressing time as a count of
nanoseconds.

Fork the implementation accordingly between Dovetail and I-pipe
configurations, depending on which pipeline flavour was detected at
init.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>