cobalt/thread: Make sure relax inband work is on a stack outliving the wakeup
As dovetail does not copy the work passed to pipeline_post_inband_work,
we must ensure that the one used in xnthread_relax lives as long as the
wakeup takes. Therefore, fold post_wakeup into the only caller so that
the stack frame of xnthread_relax is guaranteed to be used which
fulfills this requirement.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>