cobalt/thread: Pull kthread inband work onto creator stack
As dovetail does not copy the work passed to pipeline_post_inband_work,
we must ensure that the one used in xnthread_relax lives as long as the
wakeup takes. Therefore, embed the pipeline_inband_work struct needed
for signaling the kthread creator into kthread_arg struct. This lives on
the creator stack, as long as it takes to signal completion to it.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>