cobalt/posix/syscall: Handle clock_gettime[64] coming from vDSO in primary
This is needed when there is not userspace-readable clocksource
available. This can happen under Dovetail where Linux selects the
suitable clocksource and may see problems with, e.g., the TSC under x86.
When it then switches to a clocksource that require a syscall, we so
far migrate to Linux on every access.

This handler resolves that problem. It's not needed under I-pipe where
clock_gettime is not routed through the vDSO and the clocksource is
separate. Therefore, we simply never match on the related syscall
numbers under I-pipe.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>