cobalt/rtdm/fd: Fix default handler setups
2db562ad62ec had a point in avoiding EBADF for file descriptors that are
not under RTDM management. But it unfortunately got a lot of other cases
wrong along this.

This one is about returning the right error code if an RTDM driver does
not implement a particular handler. That must never be EADV because that
error would signal "not an RTDM fd, retry with Linux" which is obviously
wrong. Even worse, mmap clearly defines the right error to return in
case it is not implemented, and that is ENODEV as was returned before.

So this restores ENODEV as invalid default for mmap, uses ENOTTY for
unimplemented IOCTL and switches all others to EBADF which is a match
for select and a reasonable choice for read/write/recvmsg/sendmsg.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>