cobalt/kevent: ipipe: Check for FPU events before debug events
[backport of b8ed406edb0a26e7456acf29aefc75e01ad7f7f3]

At least on ARM, an illegal instruction event that could also mean
"breakpoint" can become an FPU event after analyzing it in
xnarch_handle_fpu_fault. Avoid registering the process as being stopped
for debugging first - this would be false alarm, and it would not be
unregistered any time soon.

This was exposed by the ptrace helper addition.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>