Bump ABI revisions
The introduction of dovetail support with the change in the realtime
clock semantics and related interface changes (vDSO usage) is an ABI
change that needs a revision bump.

Strictly spoken, PowerPC is only theoretically affected as it lacks
dovetail support, but it's simpler to bump than to safely exclude all
possibly side effects even on this arch.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>