testsuite/xeno-test: Restore timesyncd running state on smokey failure
We stop systemd-timesyncd before running the smokey testsuite and start
it again once the smokey tests are complete. As xeno-test has a "set -e"
we will stop the execution on the first error. So if smokey reports an
error systemd-timesyncd is not started again. We don't properly restore
the system state in this case.

Install a bash trap to make sure that we restart the service in all
cases if it was running on script entry.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>