Drop PowerPC architecture support
While this architecture served well in many use cases over a long time,
it's time is over, new chip chips are no longer being released.

As Xenomai is moving its core completely over Dovetail, there is no
support in Dovetail for PowerPC and that is not going to change, it's
time to drop this arch from the core in order to move forward.

This addresses the majority of artifacts, just few traces in drivers and
documentations are left for later cleanups.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>