drivers/can: add a minimal support to ethtool API
Like in the Linux kernel, the API allows to get/set device controller
parameters.
For now it only supports:
 - ETHTOOL_GDRVINFO: get general driver and device information
 - ETHTOOL_GRINGPARAM: get RX/TX ring parameters

It allows the userspace to know the number of messages managed by the
controller and how many of them are used for reception/transmission.
This information allows you to know the maximum number of messages that
sendmsg() can transmit without returning a full FIFO error.

Application programs are expected to use the new API directly via IOCTL,
since, unlike the Linux Kernel, there is currently no userspace
application like ethtool in Xenomai.

Signed-off-by: Dario Binacchi <dariobin@libero.it>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>