cobalt/thread: Keep IRQs off longer during in-band migration
This restores the pattern used before 98288382bd46: We must prevent any
interruption between queuing the inband work and suspending the current
RT thread. Otherwise, we risk that the former is consumed prior to the
thread consumption if the thread should undergo temporary remote
suspension/resumption, allowing Linux to run first. This is also in line
with the EVL core.

Fixes: 98288382bd46 ("cobalt/thread: dovetail: keep hard irqs off on transition to in-band")
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>