utils/net/rtnet: use NETMASK also with TDMA_CONFIG on master stations
Before this change, on TDMA master stations using the TDMA_CONFIG file
the NETMASK parameter was not used and netmask was assigned automatically
by [rt]ifconfig (deriving from IP class of the address assigned to the station),
both on master and on slaves.
Instead, on TDMA master stations not using the TDMA_CONFIG file, the NETMASK
parameter specified in rtnet.conf was used both for master and slaves.
This commit uniforms the NETMASK usage between configurations with and without
TDMA_CONFIG file.

Signed-off-by: Mauro Salvini <mau.salvi@tin.it>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>