cobalt/x86: Fix FPU switch during thread hardening
The current place to account for the missing call to switch_fpu_return
when returning to userland in primary mode was missing one case: the
switch-in of a thread while migrating from secondary to primary mode.
Move the logic to xnarch_switch_fpu which is also called in that case.
This also avoids having to deal with NULL xnthread_current.

Reported-by: Stefan Heber <Stefan.Heber@zwickroell.com>
Fixes: 608c23297c4d ("cobalt/x86: Account for FPU changes in kernel 5.2")
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>